// frontend/src/app/core/models/product.model.ts
export interface ProductImage {
  id: number;
  image: string;
  is_primary: boolean;
  order: number;
}

export interface Product {
  id: string;
  name: string;
  slug: string;
  description: string;
  price: string;
  discount_price?: string;
  stock: number;
  is_active: boolean;
  category: number;
  seller: number;
  images: ProductImage[];
  created_at: string;
}